This applies `PoolingAsyncValueTaskMethodBuilder` in the similar manner to what `Socket` and `Stream` do.
While it does not improve read performance itself, on low-allocation asynchronous paths I discovered this to be a contributor to GC collection and pause frequency, making it profitable.

```
BenchmarkDotNet v0.13.12, Windows 11 (10.0.22631.3296/23H2/2023Update/SunValley3)
AMD Ryzen 7 5800X, 1 CPU, 16 logical and 8 physical cores
.NET SDK 8.0.202
  [Host]     : .NET 8.0.3 (8.0.324.11423), X64 RyuJIT AVX2
  DefaultJob : .NET 8.0.3 (8.0.324.11423), X64 RyuJIT AVX2
```
Before:
| Method                 | Mean     | Error   | StdDev  | Gen0       | Allocated |
|----------------------- |---------:|--------:|--------:|-----------:|----------:|
| StreamReader           | 140.5 ms | 1.36 ms | 1.20 ms | 11500.0000 | 186.49 MB |
| Utf8StreamReader       | 174.9 ms | 2.26 ms | 2.12 ms |          - |    2.1 MB |
| Utf8TextReader         | 192.7 ms | 0.33 ms | 0.30 ms |          - |    2.1 MB |
| Utf8TextReaderToString | 192.4 ms | 1.06 ms | 0.94 ms |          - |    2.1 MB |

After:
| Method                 | Mean     | Error   | StdDev  | Gen0       | Allocated    |
|----------------------- |---------:|--------:|--------:|-----------:|-------------:|
| StreamReader           | 138.6 ms | 1.48 ms | 1.38 ms | 11666.6667 | 190967.17 KB |
| Utf8StreamReader       | 177.6 ms | 2.83 ms | 2.64 ms |          - |      1.16 KB |
| Utf8TextReader         | 197.8 ms | 0.73 ms | 0.65 ms |          - |      1.38 KB |
| Utf8TextReaderToString | 192.0 ms | 0.63 ms | 0.56 ms |          - |       1.2 KB |
|                    😅\*          |  78.87 ms | 1.371 ms | 1.216 ms |  4000.0000 |  69238.38 KB |

\*The main bottleneck especially on Windows for File IO is the latency of calls into kernel APIs. Sometimes it gets particularly bad with Windows's implementation of Overlapped. In fact, on Unix systems it is much more well-behaved despite more naive-ish implementation of blocking file read as a separate threadpool work item - the latency is way lower and this approach does not even cause issues thanks to Threadpool's ability to cope with blocked workers really well. One way to address this is increasing starting buffer size to 4KiB or even 8KiB, which U8Reader does. My theory here is modern NVME drives have pretty large memory pages internally, as well as filesystems operating on 4, 8 or even 32KiB pages themselves. This way a single file read could naturally align and let OS pull in one or multiple evenly sliced pages. I don't know whether this in fact helps, but increasing buffer size up to 8KiB has shown consistent improvement on all systems for all situations.

There is also additional cost to async depth even if valuetasks are pooled. In fact, it is so problematic that I had to rewrite `ReadToAsyncCore` and various methods forwarding to it in U8Reader entirely in order to make `U8SplitReader` (effectively a line reader variant for custom delimiters) profitable to use - it is still expensive-ish, but at least it gives reason for users to do a more memory-friendly `await foreach (var segment in file.AsU8Reader(false).Split(',')` over pre-buffering large text segments and using normal split iterator.

If the line length in your use case is substantial, or the data becomes available in small chunks, like a network stream delimited by lines, this will not help - all calls end up doing the asynchronous `await`. But if you expect the buffer to have, let's say 5 or more lines, it becomes important to avoid as many `await`s as possible even if they complete synchronously (which is why I'm also using `RandomAccess`).