@neuecc 2024-03-28
I had a problem regarding the Pool attribute.
In the end I didn't put it on,
It has allowed me to hear your great insights!

Thank you for the benchmarks as well.
I hadn't measured the performance specifically for File I/O.
I re-measured with different parameters based on your suggestions.

![image](https://github.com/Cysharp/Utf8StreamReader/assets/46207/b3be448b-177c-4ce6-ac3a-36ec9aa1c470)

Indeed, the current default (NoBuffer_Async) doesn't sho very good results.
Considering that the read buffer may not always be fully utilized in the NoBuffer case,
it might be worth preparing a separate buffer for reading, distinct from the working buffer.

The results of the microbenchmarks make it tricky to decide how to handle useAsync.