@neuecc 2024-03-28
Due to the nature of the implementation, the larger the buffer, the fewer copies and ReadAsync calls, which directly impacts performance.

![image](https://github.com/Cysharp/Utf8StreamReader/assets/46207/4e0bb43e-16a4-4c5f-9997-68357a8db341)

In serializers like MessagePack and MemoryPack, I've been using 64K buffers obtained from the ArrayPool.
So I'm inclined to think 64K might be a good choice for this as well.