@neuecc 2024-03-28
Looking back at the .NET 6 release documentation, I reviewed the following:
https://devblogs.microsoft.com/dotnet/performance-improvements-in-net-6/#io
https://devblogs.microsoft.com/dotnet/file-io-improvements-in-dotnet-6/
https://adamsitnik.com/files/Fast\_File\_IO\_with\_DOTNET\_6.pdf

Regarding Read, in the detailed explanation by Adam Sitnik, the comparison of async true/false in a Windows environment is omitted!
However, in the first article, the comparison of async:true/false being tested is included, and it can be confirmed that when true, the results are significantly slower.
It seems like they didn't want to include a table that would be unfavorable (requiring accurate explanation)...

In theory, to summarize:

* Using async I/O reduces thread pool consumption, increasing scalability
* Not using async I/O increases thread pool consumption, but it's faster in a Windows environment

I think I want to add an option to change the useAsync argument of Utf8StreamReader.
Also, let's set the buffer size to 64K.