@neon-sunset 2024-03-30
Interesting, I did not expect it to scale so well to such a large buffer size. It seems to be much more pronounced on Windows too. Reading your notes gave me some food for thought that I can use use for yet another rewrite of `U8Reader`.  There's still a lot of performance left on the table (especially in regards to GC-friendly allocation patterns) that can be gained through conditional double-buffering, adapting to read modes on the fly, detecting sync/async modes and then feeding that "stratregy" to conditionally re-allocating a handle for sync/async on Windows, and more...though that might be an overkill.

Either way, thank you and have a good day :)